Display seconds
This makes the display ambiguous!  "1:02" could mean "1 minute, two seconds"
or it could mean "1 hour, two minutes".  The only way to tell the difference
is to watch it for one second and see if it increments.  I.e., screenshots
become ambiguous, but I don't expect that this will cause much confusion
with live usage.

This stretches the amount of stuff that can fit in this small space.
In particular, the documentation for this interface[1] says that this text
"should have 4 characters or less" and "only about four [characters] can
fit in the space."  A "59:59" display requires five characters.

I think this will be mostly OK.  This uses a variable-width font on my machine.
The colon character is narrow and digits are not especially wide.

[1] https://developer.chrome.com/extensions/browserAction